moved ReadsType to Util
